+++ src/project4/CallCenterExp.scala
@@ -70,7 +70,7 @@ object CallCenterExp extends App with ProcessInteractionSimulation {
   //a caller in this simulation
   case class caller(callerNumber : Int) extends SimActor() {
     //to be set upon arrival
-    var arrivalTime : Int = 0
+    var arrivalTime : Double = 0.0
 
     //the 'script' for this actor
     actions.push("leave","usePhone","arrive")
